refactor: consolidate file endpoints into frappe/core/api/file.py
Move upload_file and download_file (with check_write_permission and
ALLOWED_MIMETYPES) from frappe/handler.py and add_attachments from
frappe/utils/file_manager.py into the existing frappe/core/api/file.py,
and mark every endpoint there @frappe.public(group="Files") with full
annotations and docstrings.

handler.py keeps plain import aliases (not lazy __getattr__) because the
deprecated bare-cmd shorthand /api/method/upload_file resolves through
that module's globals(). file_manager.py gets a lazy alias.

Also fixes a dead link: personal data download emails pointed at
frappe.utils.file_manager.download_file, which was dropped in #26575;
they now use the canonical download endpoint.
